When trying to add a gpt label to the disk that already had one, we failed.  We silently failed, alas, and this change doesn't fix that. We assumed that if we couldn't create the GPT table, then it was already properly labelled.  However, this is an invalid assumption.
So, just to be safe, we overwrite the first 10MiB and last 3ish MiB of
disk space with zeros.  This totally obliterates the old meta data
that was there so we can "start clean" when we put the gpt goo in place.

Another effect of this was that if you were taking N disks and putting
them into a zpool, and those disks already had gpt labels on them, it
would fail silently.  The GUI would should the volume, but it wasn't
there.  Sharing it didn't work, etc, etc.  Lots of bad things would
happen.  This change fixes that, but not the silent failure bit.