set default password as UNUSABLE_PASSWORD for initial_data.json (because of r7236)
Note: this is a workaround, using initial_data.json might be overkill since r7236.
We might want to simply create an admin (is_superuser=True) and password='!' in the db
instead of maintaining/loading this fixture.
For a command line case, in which creating this user is not trivial a django command
could be used for this task,