Merge r7533 from trunk:
The config database that should be backed up (and
subsequently restored in the event of update failure) is the
original config database -- not the uploaded config database.

This should prevent people from accidentally bricking boxes
when doing upgrades on broken systems, or uncaught bugs
related to updated DB schema changes.