Merge r8085:
Fix the alert subsystem so that it _actually_ reports alerts and is more robust.

The salient point was that the input provided to the log method was unicode objects; thus, by passing them through the unicode built-in, the email method no longer barfed when trying to send an alert as they were properly marshalled into string objects.