Fix regression committed via r8332 with properly configured SMTP setups.
The problem is that not configuring an SMTP host causes .connect() to fail with socket.ECONN as it tries to connect to localhost:25 -- instead of the configured mail server -- because when you omit the arguments to connect(), it tries to use whatever gets passed into the initializer, which for some odd reason is being silently filtered out by smtplib (I haven't traced down the root cause here, yet..). But yes... we'll have to fix the GUI bug with it permitting empty SMTP hosts.

I am PURPOSELY filtering out these messages so they'll only end up in the test message pane and won't blow up syslog as we are using a puny 75MB memory disk for storing things in /var today and I don't want home users to start complaining that their boxes are tipping over at home.

Identified in a support call as the customer noted that things worked when SMTP authentication wasn't specified with FreeNAS 8.0.2-RELEASE, but things were broken with a newer version.