Merge changes from trunk to 8.0.3 ------------------------------------------------------------------------ r8931 | yaberauneya | 2011-12-02 01:03:57 -0800 (Fri, 02 Dec 2011) | 6 lines
Hack in fixes for simple GUI driven cases that have been broken with AD for a long time (clock skew with PDC, domain realm join failures, etc) because I'm tired of the fact that I have to repeat myself over the phone so many times with customers and that I have to handhold people in order to ensure that they're install works when in reality the software itself was broken.

Also, call forcestop on ix-activedirectory and ix-kinit so it actually unjoins from the domain and deletes the Kerberos tickets because it was always assuming that it was in start state as opposed to stopping state.

This fixes the high ticket items that have been pissing me off for a long time with our middleware. Now, to fix the other brokenness at boot... but first... sleep... 

------------------------------------------------------------------------
r8929 | yaberauneya | 2011-12-02 00:30:00 -0800 (Fri, 02 Dec 2011) | 2 lines

Add hook for determining the current status of active directory integration from net ads status and fail if joining the domain fails.

------------------------------------------------------------------------
r8928 | yaberauneya | 2011-12-02 00:02:58 -0800 (Fri, 02 Dec 2011) | 2 lines

Add a hook for obtaining status on the current kerberos tickets on the system.