Merge r9507 from trunk.
Rework the samba services to be executed independent of one another and also to honor rc.d better.

The rc.d script reworking was submitted per PR # ports/164010, s.t. we can
now manipulate the states of the individual samba services instead of having
to restart the entire lot and kill services manually (I modified the AD and
CIFS services to follow the new world order, but I didn't change the LDAP code as I don't have the required infrastructure to test the change, yet). This
saves time when manipulating AD and CIFS now and reduces service interruption
(CIFS shares stay up and running when AD is turned off, and now AD can be on
without having CIFS enabled).