Build plugins jails with the build system.
(best sung along to http://www.youtube.com/watch?v=vVDJvrBFDDc)

"It's hacky. It's tacky. It's really, really nasty. Altogether-not-integrated-all-that-well-into-nanobsd-without-breaking-backwards-compatibility-awesome'y. This is what we do with a couple days to code before a BETA release :)!"

It works, but we'll need to revisit how this is done later. Too many things are getting stuffed into freenas-common nowadays and something similar to create_plugins_jail will need to be done to build deterministic plugins [eventually] (I foresee a build script in build/ in the not-so-immediate future and some shuffling of code needing to be done in nanobsd/common, but not now). I have an idea of what needs to be done but it needs to be documented and agreed upon later.

The jail build/install mnemonics function similar to how things are done in the base system.

They will be split up eventually. This is just a stopgap enhancement.