Disable directory name caching in samba to improve performance in some circumstances with large directories on FreeNAS
Long story short: it's busted on *BSD (Linux semantics/assumptions), smb.conf(5) says turn it off, and there's a bug for it [1] with tons of context as to why the recommendations were made to disable it on *BSD, with multiple comments from BSD long-timers/alumnus.

A more generic patch for this which will disable the support in *BSD at build time might be coming soon (delphij has an outstanding patch with BaT CCed; maintainer no ACK).

1. https://bugzilla.samba.org/show_bug.cgi?id=4715

Discussed-with: delphij (earlier version)