Improve some plugins-related descriptions
Make the fact that 'plugin path' is ultimately the location where plugin PBIs are put after installation more apparent, as the previous description was confusing.

Make it clearer that ultimately the error is that the 'plugins archive path' cannot be a subset of the 'plugins jail path' and vice versa, as the previous wording was a bit open to interpretation.

Acked-by: matto
Reviewed-by: jhixson (earlier version), dru@ (earlier version)