Merge changes from  to branches/8.2.0 ------------------------------------------------------------------------ r10562 | yaberauneya | 2012-03-08 00:06:34 -0800 (Thu, 08 Mar 2012) | 9 lines
Improve some plugins-related descriptions

Make the fact that 'plugin path' is ultimately the location where plugin PBIs are put after installation more apparent, as the previous description was confusing.

Make it clearer that ultimately the error is that the 'plugins archive path' cannot be a subset of the 'plugins jail path' and vice versa, as the previous wording was a bit open to interpretation.

Acked-by: matto
Reviewed-by: jhixson (earlier version), dru@ (earlier version)