Mute noise of log in notifier and use python logging
As a follow up of #1343 I've muted all syslogs from middleware and set
them for DEBUG.
Also I have replaced the syslog calls by python logging framework, I
thought this could benefit us in a near feature with more fine-grained
logging configuration (writing our own handlers etc). Currently the
SyslogHandler is active in django settings.py.
I know there are plans to rewrite the middleware, but while that does
not happen this could be of some use...

The reason for muting is that as already stated in #1343 I also have
heard of lot of complains about noise in /var/log/messages which are
not helpful to the user and could lead to the miss of more critical
error messages.
We could also enable /var/log/debug.log limiting it to a few hundred lines.