Merge changes from trunk to branches/8.2.0 ------------------------------------------------------------------------ r10699 | yaberauneya | 2012-03-15 09:17:06 -0700 (Thu, 15 Mar 2012) | 2 lines
This caveat in the errata section at the bottom of the release notes shouldn't be noted anymore.

------------------------------------------------------------------------
r10698 | yaberauneya | 2012-03-15 09:06:21 -0700 (Thu, 15 Mar 2012) | 62 lines

Convert GUI Upgrades from a .xz disk image payloads to .txz disk image payloads

This allows us to add upgrade verification scripts without having to
deal with the associated complexity of cracking open the disk image and
analyzing/manipulating the disk image contents, as well as ensuring
that the md(4) pseudo device is managed properly. I've done this more
than once before in a previous "life" -- it turns into a mired mess
unless certain hacky assumptions are made (which won't scale in a
project that requires production level quality, such as FreeNAS).

The verification checks added with this commit achieve the following
goals:

1. Prevent downgrading: migration downgrades don't work in the current
   design. tl;dr: sometimes people have downgraded from nightlies to
   earlier versions and their config data has gotten busted. This
   check doesn't fix importing later versions of the config database in
   older versions of software (that will require some SQL versioning fu,
   which is outside my area of expertise ;)..), but it does prevent
   simple downgrade cases.
2. Switching architectures: more than a handful of people have upgraded
   via the GUI to another architecture (x64 <-> x86) and discovered that
   their reporting system is broken, because collectd stores data in an
   architecture specific format. Similarly, one shouldn't be able to
   switch architectures (some claim that it was a feature, but it's a
   really silly design flaw), because this would brick machines in the
   future if someone switched between two mutually incompatible
   architectures (amd64 <-> mips, for example). We're not there yet,
   but that's the ultimate goal I have for FreeNAS (to have it run on
   something embedded that's not x86 based).
3. Switching projects: there was nothing preventing someone from
   switching between FreeNAS and TrueNAS, for example. Now that is
   actively prevented for various technical and business reasons
   (which I won't get into here). But this will also prevent others
   from switching between various FreeNAS 8.x derivative projects in
   the future.
4. Hardware requirements checks: some soft warnings are provided for
   cases dealing with scenarios where CPU, RAM, etc are way too low on
   the target system. These aren't hard errors because I don't want to
   break my or others' VMs, but they might become hard errors in the
   future once things are better ironed out in terms of hardware
   requirements; there's a webpage on freenas.org that states the
   hardware requirements, but they aren't barebones requirements and
   they a) aren't necessarily representative of what the requirements
   will be in the future, b) factor in software requirements for
   plugins, c) don't represent the minimum specs for all cases.

This work was done for FreeNAS to to ensure end-user POLA is maintained
when upgrading, and to reduce unnecessary support tickets and forum
posts in regard to the former item.

This mechanism will be used in the future for other things as well
(there's an additional usecase coming around the bend in the 8.3.0+
timeframe for mps end-users, dealing with the mps driver upgrade).

Support for .xz payloads will exist purely for the 8.2.0 release cycle;
all subsequent releases will switch over to .txz only payloads. Upgrading
to releases following 8.2.0 will require either upgrading via the CD
ISO or upgrading via the GUI to 8.2.0 first, then doing a GUI upgrade to
the desired final release of one's choosing (assuming that there aren't
any other upgrade prereqs).