Committing progress for transmission plugin written in django
The pre-requisite is using FastCGI to communicate with the FreeNAS GUI.
We don't have to do all the hard work, lets use a framework which can
help us with the FastCGI handling, URL routing, data persistence,
migration, form validation, etc.

We don't have to use Django, that could be PHP, Ruby, Java (Ugh!),
but I've choose it because we are all more confortable with it right now.

I've plans to write some other plugin using PHP as example.