Merge changes from trunk to 8.2.0 ------------------------------------------------------------------------ r10805 | yaberauneya | 2012-03-22 01:00:37 -0700 (Thu, 22 Mar 2012) | 8 lines
Increase the kern.cam.boot_delay from 10 to 30 seconds.

Backstory is that in some configurations with JBODs, etc it takes a really long time to probe USB media, which can result in boot failures due to device attach/probe timing issues.

These timing issues seem to be largely resolved in 9.0, BUT the testing I did with a 9.0-STABLE base was limited so this will need to be revisited later.

Do-Not-Merge (ix): not necessary