Merge changes from trunk to branches/8.2.0 ------------------------------------------------------------------------ r10821 | yaberauneya | 2012-03-22 10:45:48 -0700 (Thu, 22 Mar 2012) | 10 lines
rc.shutdown.local exists for the explicit purpose of allowing local daemons
to shut down on FreeBSD systems. Instead of running /root/save_rrds.sh in
rc.shutdown at the end where it needs to touch filesystems, it should be run
via local, where there's a better chance of it doing the right thing because of
how the services are ordered.

Reviewed-by: delphij, jpaetzel

Do-Not-Merge (ix): RCS differences won't let this merge