Hand apply changes from trunk (the logging stuff hasn't been merged here yet)
No effort was done to make this pretty because the dev merging the logging
changes will need to deal with conflicts.
------------------------------------------------------------------------
r10857 | yaberauneya | 2012-03-26 19:07:11 -0700 (Mon, 26 Mar 2012) | 4 lines

Shift the return value over by 8 so the semantics match
system(3)'s semantics and are more intuitive, as opposed
to the perl/python wait(2) semantics.