Merge changes from trunk to branches/8.2.0 ------------------------------------------------------------------------ r10902 | yaberauneya | 2012-03-29 15:56:31 -0700 (Thu, 29 Mar 2012) | 4 lines
s/service.pack/hotpatch/gi

This is what the rest of the professional world calls (IronPort, Xen, etc) it, and this is effectively what it is as they're smaller quanta of data than GUI Upgrades.

------------------------------------------------------------------------
r10901 | yaberauneya | 2012-03-29 15:39:11 -0700 (Thu, 29 Mar 2012) | 9 lines

Get rid of Service Pack dialog/windows in the GUI.

Ever since the code converged in r10894, the need for
having a separate window to point to the same exact
thing is reduced.

So let's get rid of unnecessary complexity and
confusion as this has confused users in the past.

------------------------------------------------------------------------
r10894 | yaberauneya | 2012-03-29 06:26:25 -0700 (Thu, 29 Mar 2012) | 4 lines

Make service packs work and refactor the GUI/service pack code
so it's more manageable via install_worker and so it has a
common validation and update path.