Merge r10642 from trunk:
Provide a way to develop plugins using django development server

We do that by embedding a fast cgi client based on flup which does
the same job as nginx fastcgi proxy.
I also had to monkey patch the django http server to use threads,
which is going to be default in django 1.4.