Add two fields to automatic replication so that the start time of replication can be controlled.
This is useful for scenarios like, e.g. the administrator wants
the snapshot task be done for normal office hours, but the replication
be done only in the nights.  Note that the end time for replication
is the last time for replication to *start*, for instance, a window
between 00:00 to 02:00 means a replication can be initiated inside
this window, but not end before 02:00.  In that case, the replication
will continue until it finishes.